# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TravelOS – Multi-tenant Travel ERP SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users in company
      parameters:
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /v1/users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── COMPANY DASHBOARD ─────────────────────────────────────────────────────
...
  /v1/vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List vendors
      responses:
        "200":
          description: Vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /v1/vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    delete:
      operationId: deleteVendor
      tags: [vendors]
      summary: Delete vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── CRM ───────────────────────────────────────────────────────────────────
...
  /v1/tours/destinations:
    get:
      operationId: listDestinations
      tags: [tours]
      summary: List destinations
      responses:
        "200":
          description: Destinations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Destination"
    post:
      operationId: createDestination
      tags: [tours]
      summary: Create destination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destination"
  /v1/tours/packages:
    get:
      operationId: listTourPackages
      tags: [tours]
      summary: List tour packages
      parameters:
        - name: destinationId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Tour packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TourPackage"
    post:
      operationId: createTourPackage
      tags: [tours]
      summary: Create tour package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TourPackageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourPackage"
  /v1/tours/packages/{id}:
    get:
      operationId: getTourPackage
      tags: [tours]
      summary: Get tour package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tour package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourPackage"
    patch:
      operationId: updateTourPackage
      tags: [tours]
      summary: Update tour package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TourPackageUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourPackage"
    delete:
      operationId: deleteTourPackage
      tags: [tours]
      summary: Delete tour package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── FINANCE ───────────────────────────────────────────────────────────────
...
  /v1/public/enquiry:
    post:
      operationId: submitEnquiry
      tags: [public]
      summary: Submit booking enquiry from website
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnquiryInput"
      responses:
        "201":
          description: Enquiry submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
components:
  schemas:
...
    User:
      type: object
      required: [id, email, name, role]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
        companyId:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    UserInput:
      type: object
      required: [email, name, role, password]
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
        password:
          type: string
        phone:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        isActive:
          type: boolean

    # ── Master Dashboard ───────────────────────────────────────────────────
...
    Vendor:
      type: object
      required: [id, name, contactName, phone, status]
      properties:
        id:
          type: string
        name:
          type: string
        contactName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        gstNumber:
          type: ["string", "null"]
        commissionPercent:
          type: ["number", "null"]
        totalTrips:
          type: integer
        pendingSettlement:
          type: number
        createdAt:
          type: string
    VendorInput:
      type: object
      required: [name, contactName, phone]
      properties:
        name:
          type: string
        contactName:
          type: string
        phone:
          type: string
        email:
          type: string
        gstNumber:
          type: string
        commissionPercent:
          type: number
    VendorUpdate:
      type: object
      properties:
        contactName:
          type: string
        phone:
          type: string
        status:
          type: string
        commissionPercent:
          type: number

    # ── Lead / CRM ─────────────────────────────────────────────────────────
    Lead:
      type: object
      required: [id, name, phone, source, status]
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        source:
          type: string
          enum: [website, whatsapp, phone, email, walk_in, referral]
        status:
          type: string
          enum: [new, contacted, qualified, quotation_sent, won, lost]
        assignedTo:
          type: ["string", "null"]
        travelDate:
          type: ["string", "null"]
        destination:
          type: ["string", "null"]
        pax:
          type: ["integer", "null"]
        budget:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        followUpDate:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Destination:
      type: object
      required: [id, name, country]
      properties:
        id:
          type: string
        name:
          type: string
        state:
          type: ["string", "null"]
        country:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        totalPackages:
          type: integer
    DestinationInput:
      type: object
      required: [name, country]
      properties:
        name:
          type: string
        state:
          type: string
        country:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string

    # ── Tour Package ───────────────────────────────────────────────────────
    TourPackage:
      type: object
      required: [id, title, duration, price, destinationId]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        duration:
          type: integer
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        destinationId:
          type: string
        destinationName:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        inclusions:
          type: array
          items:
            type: string
        exclusions:
          type: array
          items:
            type: string
        highlights:
          type: array
          items:
            type: string
        rating:
          type: ["number", "null"]
        totalBookings:
          type: integer
        isActive:
          type: boolean
        packageType:
          type: string
          enum: [adventure, pilgrimage, luxury, corporate, honeymoon, family, group]
    TourPackageInput:
      type: object
      required: [title, duration, price, destinationId, packageType]
      properties:
        title:
          type: string
        description:
          type: string
        duration:
          type: integer
        price:
          type: number
        originalPrice:
          type: number
        destinationId:
          type: string
        packageType:
          type: string
        inclusions:
          type: array
          items:
            type: string
        exclusions:
          type: array
          items:
            type: string
        highlights:
          type: array
          items:
            type: string
    TourPackageUpdate:
      type: object
      properties:
        title:
          type: string
        price:
          type: number
        isActive:
          type: boolean
        description:
          type: string

    # ── Finance ────────────────────────────────────────────────────────────
...
    EnquiryInput:
      type: object
      required: [name, phone, destination, travelDate]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        destination:
          type: string
        travelDate:
          type: string
        pax:
          type: integer
        budget:
          type: number
        message:
          type: string